+++ src/main/java/com/scyllacore/dumpWeb/commonModule/db/mapper/manage/Step6ForVehicleManageMileageViewerMapper.java
@@ -11,8 +11,8 @@ import java.util.List;
 public interface Step6ForVehicleManageMileageViewerMapper {
     List<MileageDTO.Response> selectMileageListByOption(PageDTO option);
 
-    int countMileageListByOption(PageDTO option);
+    long countMileageListByOption(PageDTO option);
 
-    int updateMileagePaymentChk(MileageSearchOptionDTO.Request option);
+    long updateMileagePaymentChk(MileageSearchOptionDTO.Request option);
 
 }
